Crash safe slave when using GTID and MTS
Summary:
Gtid information is stored in a transaction table
(slave_gtid_info). The table contains three columns

  Id  | Database name  | Last_gtid

Every database contains a row in the table storing the last gtid
in the replication stream for that database. On every commit by the
sql_thread or slave worker thread in case of MTS, the corresponding
row in the table slave_gtid_info is updated using XA. The table schema
slave_worker_info is changed to include last_gtid executed by that worker.
This extra information helps in finding out the group where worker stopped
in case of an error.

Every group gtid is checked with the last_gtid seen for that database to
decide whether it should be executed or skipped in which case only a empty
group of that gtid is printed in the binary log. Empty group in the binlog looks like
this

  SET @@session.GTID_NEXT='';
  BEGIN;
  // Followup of this diff (either wormhole comments or the whole event)
  COMMIT;

Current setting for crash safe slave with gtid and mts requires

  --relay_log_info_repository='TABLE'
  (currently this won't work with MTS, a restart with relay-log-recovery=0 followed by
  START SLAVE UNTIL SQL_AFTER_MTS_GAPS is required)
  --relay_log_recovery=1
  --gtid_mode=ON

without --relay_log_recovery=1 a transaction can get lost if the relay log ends with GTID
event. With this transactional table the new crash safe slave parameters with gtid and
MTS requires.

  --relay_log_recovery=1
  --gtid_mode=ON

This avoids the need for START SLAVE UNTIL SQL_AFTER_MTS_GAPS command and sync_relay_log=1.
The disadvantage is to copy binlogs again from master if slave is severely lagged.

A slave running with sync_binlog=0 may have missing binlog events in case
of a crash, but the events are committed in the storage engine. With new gtid crash
safe functionality using a transaction table slave_gtid_info, the missing events
can be detected. This diff also adds the functionality to write the missing events into
the binlog, only when using statement based replication.

coordinator thread doesn't need to store gtid_info in MTS mode, the same
work is done by worker threads and worker threads flushes the
information to mysql.slave_gtid_info table on xid commit.

Fixed a bug in in insert_dynamic_set() where it fails to de-duplicate
leading to increased size of dynamic array rli->gtid_infos.

Test Plan:
not yet implemented

The original code can be found
here:
https://github.com/facebook/mysql-5.6
branch:
webscalesql-5.6.21.55
commit:
82fa6d92da1f6994a06e450a7f269b0394e8fe9e