tokudb_dir_per_db option is introduced.
The option is on by default. If this option is on then newly created or
renamed tokudb table files will be placed into directories which correspond
to database names in data directory otherwise the files will be in root
data directory.

Some table can be converted from one directory layout to another with "rename"
operation. When some table is renamed the table files will be renamed into
directory layout which corresponds to tokudb_dir_per_db option.