lp-1657737: Print locking transaction which cause "Lock wait timeout" errors into log file
Please don't pay attention on test and formatting.

Also there are several places in innodb where lock-wait-timeout error is
returned, these places needs to be checked. Apart from this there must be some
check if PFS is enabled both for compile and run times, and if it's not enabled,
then just output the current queries for requested and blocking transactions.

And I would add that the current implementation has some limitations. The first
limitation is about the amount of blocking transactions. Currently the
information about only one transaction is printed out, but there can be several
blocking transactions. And the second limitation - there can be the case when
blocking transaction has been already finished at the moment where we try to get
it's information from PFS or some other source. I don't see the way to block it
until the error processin is finished(and I am not sure this is reasonable as
it can unfluence performance).

The general question for me is how to pass data from storage engine to audit
plugin. We can't invoke audit plugin hooks directly because some storage
engines, for example innodb, can be used in "embedded" build, but audit plugin
can't be built for "embedded" version. That's why I changed "handler" API to
add the ability to insert some hooks in storage engines. Such solution allow us
to implement engine-specific audit plugin events, but... But it changes the API,
and I am not sure if the above is good enough reason for this. And there might
be other ways to do the same but without the API changing which I missed.

And the other question which bothers me. Currently to execute query inside the
plugin the new thread is created. The general reason why we execute queries in
a separate thread is that PFS has it's own thread id which differs from process
list id. There is map from PFS thread id to process list thread id inside of
PFS code. And when a query is executed with microservices API the new process
list id is allocated for the same THD, and the old process list id is replaced
with the new one in the map. And that's why we can't find the statements history
in PFS tables for requested lock transaction as it's process list id was
replaced by microservices API. It looks like the new thread is the only way
to execute queries inside of plugin if we want to preserve THD's process
list id.