The tests for the previous commit were added. This is temporary commit, so just squash it for final commit.
There are three tests:

bug89205.test
The more complete test which covers both left and right page splits. The test
passes with the patch and does not pass without one. It's supposed to be the
general test for PR. Also the test can be used as a basis for the non-XA test,
which, more likely, will hang, because the current patch is only for XA
transactions. Just remove all XA commands and use "BEGIN, COMMIT" instead.
If non-XA test failes then new bug report is required.

bug89205-right-split-passed.test
The original test which covers only right split case. The test behaviour is the
same as for the previous test. This test was committed to compare it with the
following one.

bug89205-right-split-failed.test
This test looks very simiar as the previous one, the only difference is in
connections sequence. Use diff with the previous test to undestand the
difference. But this difference leads to "Lock wait timeout". This case needs
additional research and, maybe, bug report.