The changes are necessary to 'open/create', 'insert/update' and 'select' functionality.
The general issue is that native partitioning initialization sequence is not
the same as for non-native partitioning. Some functions are invoked before
partition info is filled in, and, as a result, we can't initialize the array
of partition file handlers. That is why the array initialization was moved to
"open/create" functions, they are invoked when all necessary information is
gathered from statement or *.frm file.

The another issue was using the correct mem_root object for file handlers
allocation. The initial code uses thd->mem_root, what is wrong, because it
will be deallocated after query execution, and it worked only because
thd->mem_root object was replaced with share->mem_root just before file handlers
initialization. The current code uses share->mem_root explicitly.