PS-3961: Improve fix for PS-255: print blocking query when possible
The test is unstable. When '--send' is used to model the case when blocking
query has not yet been sent result to the client, and it's possible to extract
blocking query string in the cheap, sometimes another transaction
executes it's query before the table is locked, and the test fails. I would
suggest to use existing or insert new debug sync point.