Fix bug 90351 / PS-3951 (Fix decreasing status counters)
Status counters were cleaned in THD::release_resources, before an
optional thread cleanup.  The cleanup could possibly perform a
rollback, or other counter-increasing operations.  In this case, these
additions to the thread local counters were lost during the
destruction of the thread, but were visible for a short time, before
the thread was removed from the global thread list.

This change avoids this issue by moving the global counter update to
the end of the release_resources method, ensuring that possible
counter-changing operation happen before it.