Non-inplace alter partition is implemented.
Non-inplace alter partition is implemented as following:

1) New table is created
2) The data from the old table is compied to the new table

There are at least two issues:

1) When data is copied to the new table during "reorganize table" there can be
rows which do not suite to the partitions in the new table. To solve this
issue the new function to handler interface is added to ask handler if copying
must be continued on the corresponding error or not.

2) When some partition is deleted it's expected that all data in this partition
must be deleted. But when rows are copied to the new table it's impossible to
detect if some row belong to the deleted partitions or not, that is why data
from dropped partitions will be copied to the new table. This is not not fixed
as the copying code is common and in the server. And there is actually no need
to fix it because the issue can be solved implementing inplace alter partition.