Bug 93250 fix.
The result of tc_log->commit() is ignored in trans_commit_stmt() and
trans_commit_implicit(). If there is error on binlog commit, trans_commit_stmt()
and trans_commit_implicit() return "ok", what leads to the attemt to set "OK"
status in diagnostics area when the status was previously set to "Error" in
MYSQL_BIN_LOG::commit(). It does not matter for release build, because
Diagnostics_area::set_ok_status() does not change the status if the current
status is "Error". For debug build the assert checks if the status is not set,
and this cause crash.

The fix returns error from trans_commit_implicit() and trans_commit_stmt() if
tc_log->commit() return error.